Fix for possible problem with symbolic links
In case of the following setup

dir
 |_ a.c
 |_ b.c
 |_ c.c -> ../some_dir/some_other_dir/h.c

a.c, b.c and h.c will be included into the report because
revisor will resolve all symbolic links for all files in the report

Problem appear when link will be changed to point to another file
but in the report we still have a.c, b.c and h.c
Next time when you do CI build, revisor will check md5 checksum
for a.c, b.c and h.c, but they will be the same, and revisor will not notice
change. So re-build will be not triggered

Internally this problem caused by using realpath function for path normalization
Idea is that we want to have clean path's without patterns like this
bla/../../bla/bla.c and we use realpath function to do that.
But as side affect symbolic links are getting resolved and we are getting
path to the actual file instead of symbolics link in the report
Solution - normalize path by ourselves